+++ ui/src/components/menu/SideNavMenu.vue
@@ -82,25 +82,6 @@
             </v-col>
           </v-btn>
         </v-col>
-        <v-col v-if="isPersonalTribe || settings.allowStories" cols="3" md="12" :class="mobile ? 'py-0 px-0' : tablet ? 'py-4 px-0' : 'py-1'">
-          <!-- TODO: connect timeline -->
-          <v-btn @click="setActive(TIMELINE)" light :fab="mobile" text>
-          <!-- <v-btn @click="setDialog('coming-soon')" light :fab="mobile" text> -->
-            <v-col class="pa-0" :cols="mobile ? '12' : '2'">
-              <TimelineIcon
-                :size="tablet ? 'x-large' : 'medium'"
-                :color="activeComponent === TIMELINE? 'red' : 'black'"
-              />
-            </v-col>
-            <v-col class="py-0" v-if="!mobile && !isOverflowing">
-              <span
-                ref="text"
-                :style="activeComponent === TIMELINE? 'color:#B02425;' : 'black'"
-                class="ml-2 nav-label subtitle-1"
-              >{{ t(TIMELINE)}}</span>
-            </v-col>
-          </v-btn>
-        </v-col>
         <v-col cols="3" md="12" v-if="showWhakapapa && (isPersonalTribe || settings.allowWhakapapaViews)" :class="mobile ? 'py-0 px-0' : tablet ? 'py-4 px-0' : 'py-1'">
           <v-btn @click="setActive(WHAKAPAPA)" light :fab="mobile" text>
             <v-col class="pa-0" :cols="mobile ? '12' : '2'">
@@ -147,14 +128,12 @@ import ArchiveIcon from '@/components/button/ArchiveIcon.vue'
 import WhakapapaIcon from '@/components/button/WhakapapaIcon.vue'
 import PersonListIcon from '@/components/button/PersonListIcon.vue'
 import UserIcon from '@/components/button/UserIcon.vue'
-import TimelineIcon from '@/components/button/TimelineIcon.vue'
 import RegisterButton from '@/components/button/RegisterButton.vue'
 
 import Avatar from '@/components/Avatar.vue'
 import { mapGetters, mapActions } from 'vuex'
 
 const PROFILE = 'profile'
-const TIMELINE = 'timeline'
 const ARCHIVE = 'archive'
 const WHAKAPAPA = 'whakapapa'
 const PERSON = 'person'
@@ -175,7 +154,6 @@ export default {
       // magic strings
       PROFILE,
       ARCHIVE,
-      TIMELINE,
       WHAKAPAPA,
       PERSON,
 
@@ -350,7 +328,6 @@ export default {
     WhakapapaIcon,
     PersonListIcon,
     UserIcon,
-    TimelineIcon,
     Avatar,
     RegisterButton
   }